Ik heb een BuyGUI's systeem gemaakt, en heb een systeem voor granaten gemaakt. Verder heb ik gewerkt aan het toevoegen van geweren die in kunnen zoomen. En natuurlijk nog wat andere kleine dingetjes

+++ PWS-Project/src/com/daan/pws/weapon/DamagePattern.java
@@ -2,7 +2,7 @@ package com.daan.pws.weapon;
 
 public abstract class DamagePattern {
 
-	public abstract double onBulletHit(PlayerHeight height);
+	public abstract int onBulletHit(PlayerHeight height);
 
 	public static enum PlayerHeight {
